refactor(dashboard): extract shared formatting utils to lib/format.ts
Four utility functions were copy-pasted across 6 dashboard pages with
minor inconsistencies between implementations:

- formatTimeAgo: defined 5 times (different signatures: Date|null vs string)
- formatLastSeen: defined 3 times (identical)
- getStatusColor: defined 3 times (rig-groups and rigs/[id] missing REBOOTING case)
- getSeverityColor: defined 2 times (identical)

Created apps/dashboard/src/lib/format.ts with unified, well-typed versions
of all four functions. formatTimeAgo now accepts string | Date | null for
full flexibility. getStatusColor picks up the REBOOTING case from rigs/page.tsx
that two other copies were missing.

Removed local definitions from all six consumer files:
- app/page.tsx (-40 lines)
- app/alerts/page.tsx (-12 lines)
- app/alerts/rules/page.tsx (-12 lines)
- app/rigs/page.tsx (-31 lines)
- app/rigs/[id]/page.tsx (-21 lines)
- app/rig-groups/page.tsx (-8 lines)

No behavior change. Build verified clean.
