+++ apps/dashboard/src/app/alerts/page.tsx
@@ -4,6 +4,7 @@ import { useEffect, useState, useCallback } from 'react';
 import Link from 'next/link';
 import { SkeletonAlertItem } from '../../components/Skeleton';
 import { getApiUrl } from '../../lib/api';
+import { formatTimeAgo } from '../../lib/format';
 
 interface Alert {
   id: string;
@@ -166,19 +167,6 @@ export default function AlertsPage() {
           </svg>
         );
     }
-  }
-
-  function formatTimeAgo(dateString: string) {
-    const date = new Date(dateString);
-    const diff = Date.now() - date.getTime();
-    const secs = Math.floor(diff / 1000);
-    if (secs < 60) return 'Just now';
-    const mins = Math.floor(secs / 60);
-    if (mins < 60) return `${mins}m ago`;
-    const hours = Math.floor(mins / 60);
-    if (hours < 24) return `${hours}h ago`;
-    const days = Math.floor(hours / 24);
-    return `${days}d ago`;
   }
 
   const unreadCount = alerts.filter(a => !a.read).length;
