+++ apps/dashboard/src/app/alerts/rules/page.tsx
@@ -3,6 +3,7 @@
 import { useEffect, useState, useCallback } from 'react';
 import Link from 'next/link';
 import { getApiUrl, getCsrfToken } from '../../../lib/api';
+import { formatTimeAgo } from '../../../lib/format';
 
 type AlertCondition =
   | 'TEMP_ABOVE'
@@ -45,19 +46,6 @@ function getConditionLabel(condition: AlertCondition): string {
 
 function getConditionUnit(condition: AlertCondition): string {
   return CONDITIONS.find((c) => c.value === condition)?.unit ?? '';
-}
-
-function formatTimeAgo(dateString: string) {
-  const date = new Date(dateString);
-  const diff = Date.now() - date.getTime();
-  const secs = Math.floor(diff / 1000);
-  if (secs < 60) return 'Just now';
-  const mins = Math.floor(secs / 60);
-  if (mins < 60) return `${mins}m ago`;
-  const hours = Math.floor(mins / 60);
-  if (hours < 24) return `${hours}h ago`;
-  const days = Math.floor(hours / 24);
-  return `${days}d ago`;
 }
 
 const defaultForm = {
