+++ apps/dashboard/src/app/page.tsx
@@ -4,6 +4,7 @@ import { useEffect, useState, useCallback } from 'react';
 import Link from 'next/link';
 import { useWebSocket } from '../hooks/useWebSocket';
 import { getApiUrl } from '../lib/api';
+import { formatTimeAgo, formatLastSeen, getStatusColor, getSeverityColor } from '../lib/format';
 
 interface GPU {
   id: string;
@@ -266,45 +267,6 @@ export default function DashboardPage() {
       if (intervalId) clearInterval(intervalId);
     };
   }, [autoRefresh, isConnected, fetchRigs, fetchAlerts]);
-
-  function getStatusColor(status: string) {
-    switch (status) {
-      case 'ONLINE': return 'bg-green-500';
-      case 'WARNING': return 'bg-yellow-500';
-      case 'ERROR': return 'bg-red-500';
-      default: return 'bg-slate-500';
-    }
-  }
-
-  function formatLastSeen(lastSeen: string | null) {
-    if (!lastSeen) return 'Never';
-    const diff = Date.now() - new Date(lastSeen).getTime();
-    const mins = Math.floor(diff / 60000);
-    if (mins < 1) return 'Just now';
-    if (mins < 60) return `${mins}m ago`;
-    const hours = Math.floor(mins / 60);
-    if (hours < 24) return `${hours}h ago`;
-    return `${Math.floor(hours / 24)}d ago`;
-  }
-
-  function getSeverityColor(severity: string) {
-    switch (severity) {
-      case 'CRITICAL': return 'bg-red-500/20 text-red-400 border-red-500/30';
-      case 'ERROR': return 'bg-red-500/20 text-red-400 border-red-500/30';
-      case 'WARNING': return 'bg-yellow-500/20 text-yellow-400 border-yellow-500/30';
-      default: return 'bg-blue-500/20 text-blue-400 border-blue-500/30';
-    }
-  }
-
-  function formatTimeAgo(dateString: string) {
-    const diff = Date.now() - new Date(dateString).getTime();
-    const mins = Math.floor(diff / 60000);
-    if (mins < 1) return 'Just now';
-    if (mins < 60) return `${mins}m ago`;
-    const hours = Math.floor(mins / 60);
-    if (hours < 24) return `${hours}h ago`;
-    return `${Math.floor(hours / 24)}d ago`;
-  }
 
   // Group rigs by coin
   const rigsByCoin = rigs.reduce((acc, rig) => {
