+++ apps/dashboard/src/app/rigs/page.tsx
@@ -5,6 +5,7 @@ import Link from 'next/link';
 import { useWebSocket } from '../../hooks/useWebSocket';
 import { SkeletonRigCard } from '../../components/Skeleton';
 import { getApiUrl } from '../../lib/api';
+import { formatTimeAgo, getStatusColor, formatLastSeen } from '../../lib/format';
 
 interface GPU {
   id: string;
@@ -170,15 +171,6 @@ export default function RigsPage() {
     } catch (error) {
       console.error('Failed to fetch rig groups:', error);
     }
-  }
-
-  function formatTimeAgo(date: Date | null) {
-    if (!date) return '';
-    const diff = Date.now() - date.getTime();
-    const secs = Math.floor(diff / 1000);
-    if (secs < 60) return `${secs}s ago`;
-    const mins = Math.floor(secs / 60);
-    return `${mins}m ago`;
   }
 
   async function handleDelete(e: React.MouseEvent, id: string, name: string) {
@@ -394,16 +386,6 @@ export default function RigsPage() {
     }
   }
 
-  function getStatusColor(status: string) {
-    switch (status) {
-      case 'ONLINE': return 'bg-green-500';
-      case 'WARNING': return 'bg-yellow-500';
-      case 'ERROR': return 'bg-red-500';
-      case 'REBOOTING': return 'bg-blue-500';
-      default: return 'bg-slate-500';
-    }
-  }
-
   function getStatusBadge(status: string) {
     switch (status) {
       case 'ONLINE': return 'bg-green-500/10 text-green-400 ring-1 ring-green-500/30';
@@ -427,17 +409,6 @@ export default function RigsPage() {
   function getMaxTemp(gpus: GPU[]) {
     const temps = gpus.map((g) => g.temperature).filter((t): t is number => t !== null);
     return temps.length > 0 ? Math.max(...temps) : null;
-  }
-
-  function formatLastSeen(lastSeen: string | null) {
-    if (!lastSeen) return 'Never';
-    const diff = Date.now() - new Date(lastSeen).getTime();
-    const mins = Math.floor(diff / 60000);
-    if (mins < 1) return 'Just now';
-    if (mins < 60) return `${mins}m ago`;
-    const hours = Math.floor(mins / 60);
-    if (hours < 24) return `${hours}h ago`;
-    return `${Math.floor(hours / 24)}d ago`;
   }
 
   const filteredRigs = rigs.filter(rig => {
